pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-coding-agent"
version = "1.0.0"
description = "A terminal-based AI coding agent with Perceive-Reason-Act-Learn loop"
authors = [{name = "AI Agent Developer", email = "dev@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "gitpython>=3.1.0",
    "pydantic>=2.0.0",
    "aiofiles>=23.0.0",
    "asyncio>=3.4.3",
    "sqlite3",
    "python-dotenv>=1.0.0",
    "prompt-toolkit>=3.0.0",
    "colorama>=0.4.6",
    "typing-extensions>=4.0.0",
    "pathlib",
    "subprocess",
    "json",
    "yaml",
    "requests>=2.31.0",
    "tiktoken>=0.5.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0", 
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0"
]

[project.scripts]
agent = "ai_coding_agent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88